refactor: create package for report app's views
PT-1878

Move different type of report views to new view files:
- admin views
- CSV views
- reports API views

Rename the report views, move them under a new views folder and reexport
them from the views package (using __init__.py).
